requirements.txt
streamlit==0.89.0
Pillow==8.2.0
mysql-connector-python==8.0.27
google-cloud-storage==1.42.3
